+++ src/no/runsafe/dergons/DergonHandler.java
@@ -1,6 +1,7 @@
 package no.runsafe.dergons;
 
 import net.minecraft.server.v1_8_R3.*;
+import no.runsafe.dergons.event.*;
 import no.runsafe.framework.api.*;
 import no.runsafe.framework.api.event.plugin.IConfigurationChanged;
 import no.runsafe.framework.api.event.plugin.IPluginEnabled;
